Formally deprecate #use_clean_gemset
