+++ lib/aruba/api.rb
@@ -16,7 +16,6 @@ require 'aruba/api/environment'
 require 'aruba/api/filesystem'
 require 'aruba/api/text'
 require 'aruba/api/bundler'
-require 'aruba/api/rvm'
 
 Aruba.platform.require_matching_files('../matchers/**/*.rb', __FILE__)
 
@@ -29,7 +28,6 @@ module Aruba
     include Aruba::Api::Environment
     include Aruba::Api::Filesystem
     include Aruba::Api::Bundler
-    include Aruba::Api::Rvm
     include Aruba::Api::Deprecated
     include Aruba::Api::Text
   end
